Refactor Stats data loading and worker plumbing
